+++ crate/cli/src/actions/elliptic_curves/keys/create_key_pair.rs
@@ -75,8 +75,8 @@ pub struct CreateKeyPairAction {
     #[clap(long = "tag", short = 't', value_name = "TAG")]
     tags: Vec<String>,
 
-    /// The unique id of the private key; a unique id based
-    /// on the key material is generated if not specified.
+    /// The unique id of the private key; a random uuid
+    /// is generated if not specified.
     #[clap(required = false)]
     private_key_id: Option<String>,
 }