+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [4.19.1] - 2024-10-11
+
+### 🚀 Features
+
+- Client `ckms`: merge attributes handling (set/get/delete) under `attributes` subcommand ([#329](https://github.com/Cosmian/kms/pull/329))
+
+### 🐛 Bug Fixes
+
+- Guard on size of ciphertexts for BulkData ([#330](https://github.com/Cosmian/kms/pull/330))
+- KMIP Attributes: fix deletion on Links and Vendor Attributes ([#329](https://github.com/Cosmian/kms/pull/329))
+
 ## [4.19.0] - 2024-10-09
 
 ### 🚀 Features